requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.1.2
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycryptodome==3.15.0
Werkzeug==2.1.2
zipp==3.8.0
